client/requirements.txt
pyyaml
requests
httpx
schedule